--- repository/src/main/resources/data.sql
@@ -1,8 +1,0 @@
-insert into user(email, password, nickname, name. phone_number, introduction, sex, website, profile_image) values("test1@test.com", "testPassword1", "testNickname1", "testName1", "010-0000-0001", "MALE", "testWebsite1.com", "testProfileImage1");
-insert into user(email, password, nickname, name. phone_number, introduction, sex, website, profile_image) values("test2@test.com", "testPassword2", "testNickname2", "testName2", "010-0000-0002", "FEMALE", "testWebsite2.com", "testProfileImage2");
-insert into user(email, password, nickname, name. phone_number, introduction, sex, website, profile_image) values("test3@test.com", "testPassword3", "testNickname3", "testName3", "010-0000-0003", "MALE", "testWebsite3.com", "testProfileImage3");
-insert into user(email, password, nickname, name. phone_number, introduction, sex, website, profile_image) values("test4@test.com", "testPassword4", "testNickname4", "testName4", "010-0000-0004", "FEMALE", "testWebsite4.com", "testProfileImage4");
-
-insert into follow(follower_id, followee_id) values(1L, 2L);
-insert into follow(follower_id, followee_id) values(1L, 3L);
-insert into follow(follower_id, followee_id) values(1L, 4L);
